Changed the structure of the project
