+++ lib/midjourney_client.dart
@@ -1,5 +1,5 @@
 library midjourney_client;
 
-export 'src/core/model/midjourney/midjourney_message.dart';
+export 'src/core/midjourney/midjourney.dart';
+export 'src/core/midjourney/model/midjourney_message.dart';
 export 'src/core/utils/logger.dart' hide MLogger;
-export 'src/midjourney_client.dart';
