+++ lib/src/core/discord/model/discord_message.dart
