+++ lib/src/core/discord/model/discord_ws.dart
@@ -1,3 +1,4 @@
+/// Base discord model 
 sealed class DiscordWs {
   const DiscordWs();
