+++ lib/src/core/discord/model/interaction.dart
