+++ lib/src/core/midjourney/model/midjourney_config.dart
