+++ lib/src/core/midjourney/model/midjourney_message.dart
